+++ src/axom/quest/io/STLReader.hpp
@@ -1,5 +1,6 @@
-// Copyright (c) 2017-2025, Lawrence Livermore National Security, LLC and
-// other Axom Project Developers. See the top-level LICENSE file for details.
+// Copyright (c) Lawrence Livermore National Security, LLC and other
+// Axom Project Contributors. See top-level LICENSE and COPYRIGHT
+// files for dates and other details.
 //
 // SPDX-License-Identifier: (BSD-3-Clause)
 
@@ -51,13 +52,13 @@ public:
    * \brief Returns the number of nodes of the surface mesh.
    * \return numNodes the number of nodes.
    */
-  int getNumNodes() const { return m_num_nodes; };
+  int getNumNodes() const { return static_cast<int>(m_num_nodes); };
 
   /*!
    * \brief Returns the number of faces of the surface mesh.
    * \return numFaces the number of faces.
    */
-  int getNumFaces() const { return m_num_faces; };
+  int getNumFaces() const { return static_cast<int>(m_num_faces); };
 
   /*!
    * \brief Clears all internal data-structures
